Add StatsContainer support
Adds support for the docker 1.5 stats container api which sends
a stream of container stats every second.  The API takes a
StatsContainerOptions struct container the container name and a
channel to read stats.  The function will write continously to
the channel and drop an messages if the receiver is slow.